requirements.txt
flask
tensorflow # Si estas en Mac puede que debas usar tensorflow-macos
pillow
numpy
